see #160 : add agent id init or get
