+++ api/src/core/adapters/hal/getHalSoftware.ts
@@ -60,7 +60,7 @@ export async function fetchHalSoftwareById(halDocid: string): Promise<HalRawSoft
 
 export async function fetchHalSoftwares(): Promise<Array<HalRawSoftware>> {
     // Filter only software who have an swhidId to filter clean data on https://hal.science, TODO remove and set it as an option to be generic
-    const url = `https://api.archives-ouvertes.fr/search/?q=docType_s:SOFTWARE&rows=10000&fl=${halSoftwareFieldsToReturnAsString}&&fq=swhidId_s:["" TO *]`;
+    const url = `https://api.archives-ouvertes.fr/search/?q=docType_s:SOFTWARE&rows=10000&fl=${halSoftwareFieldsToReturnAsString}&fq=swhidId_s:["" TO *]`;
 
     const res = await fetch(url).catch((err) => {
         console.error(err);
