--- api/src/core/usecases/feedDB.ts
@@ -1,20 +1,0 @@
-import { DbApiV2 } from "../ports/DbApiV2";
-import HAL from "../adapters/hal"
-import { HalRawSoftwareToSoftwareForm } from "../adapters/hal/halRawSoftware";
-
-export const feedDBfromHAL : any = (dbApi : DbApiV2)  => {
-    return async () => {
-        const softwares = await HAL.software.getAll();
-
-        return softwares.map(async software => {
-            const newSoft = HalRawSoftwareToSoftwareForm(software);
-            const soft = await dbApi.software.getByName(newSoft.softwareName);
-            if (soft) {
-                return Promise.resolve(soft.id);
-            } else {
-                console.debug('Importing HAL : ', software.docid, newSoft.softwareName);
-                return dbApi.software.create({ formData: newSoft, externalDataOrigin: 'HAL', agentId: undefined });
-            }
-        });        
-    }
-}
