+++ deployments/docker-compose-example/.env.sample
@@ -13,6 +13,7 @@ SILL_API_PORT=3084
 SILL_IS_DEV_ENVIRONNEMENT=true
 SILL_EXTERNAL_SOFTWARE_DATA_ORIGIN=wikidata
 INITSOFTFROMSOURCE=false
+BOTAGENTEMAIL=contact@sill.codegouv.fr
 
 DATABASE_URL=postgresql://sill:pg_password@localhost:5432/sill
 POSTGRES_DB=sill
